fix key

+++ src/main/java/com/luna/redis/util/RedisKeyUtil.java
@@ -1,9 +1,6 @@
 package com.luna.redis.util;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -26,13 +23,13 @@ public class RedisKeyUtil {
 
     /**
      * 移动key到指定DB
-     * 
+     *
      * @param key
      * @param dbIndex
      * @return
      */
     public boolean move(String key, final int dbIndex) {
-        return redisTemplate.move(key, dbIndex);
+        return Boolean.TRUE.equals(redisTemplate.move(key, dbIndex));
     }
 
     /**
@@ -40,6 +37,9 @@ public class RedisKeyUtil {
      */
     public void flushDb() {
         RedisConnectionFactory connectionFactory = redisTemplate.getConnectionFactory();
+        if (connectionFactory == null) {
+            return;
+        }
         RedisConnection connection = connectionFactory.getConnection();
         connection.flushDb();
     }
@@ -49,6 +49,9 @@ public class RedisKeyUtil {
      */
     public void flushAll() {
         RedisConnectionFactory connectionFactory = redisTemplate.getConnectionFactory();
+        if (connectionFactory == null) {
+            return;
+        }
         RedisConnection connection = connectionFactory.getConnection();
         connection.flushAll();
     }
@@ -56,37 +59,27 @@ public class RedisKeyUtil {
     /**
      * 指定缓存失效时间
      *
-     * @param key 键
-     * @param time 时间(秒)
+     * @param key      键
+     * @param time     时间(秒)
      * @param timeUnit
      * @return
      */
     public boolean expire(String key, long time, TimeUnit timeUnit) {
         if (timeUnit != null) {
-            return redisTemplate.expire(key, time, timeUnit);
+            return Boolean.TRUE.equals(redisTemplate.expire(key, time, timeUnit));
         }
-        return redisTemplate.expire(key, time, TimeUnit.SECONDS);
+        return Boolean.TRUE.equals(redisTemplate.expire(key, time, TimeUnit.SECONDS));
     }
 
     /**
      * 指定缓存失效时间
      *
-     * @param key 键
+     * @param key  键
      * @param time 时间(秒)
      * @return
      */
     public boolean expire(String key, long time) {
         return expire(key, time, null);
-    }
-
-    /**
-     * 返回模糊匹配的key
-     * 
-     * @param key List<String>
-     * @return
-     */
-    public Set<String> getKeysWithList(String key) {
-        return redisTemplate.keys(key);
     }
 
     /**
@@ -99,7 +92,7 @@ public class RedisKeyUtil {
 
     /**
      * 返回模糊匹配的key 不重复
-     * 
+     *
      * @param key Set<String>
      * @return
      */
@@ -113,7 +106,7 @@ public class RedisKeyUtil {
      * @param key 键 不能为null
      * @return 时间(秒) 返回0代表为永久有效
      */
-    public long getExpire(String key, TimeUnit timeUnit) {
+    public Long getExpire(String key, TimeUnit timeUnit) {
         if (timeUnit == null) {
             return redisTemplate.getExpire(key, TimeUnit.SECONDS);
         }
@@ -122,7 +115,7 @@ public class RedisKeyUtil {
 
     /**
      * 根据key 获取过期时间
-     * 
+     *
      * @param key
      * @return
      */
@@ -137,12 +130,12 @@ public class RedisKeyUtil {
      * @return true 存在 false不存在
      */
     public boolean hasKey(String key) {
-        return redisTemplate.hasKey(key);
+        return Boolean.TRUE.equals(redisTemplate.hasKey(key));
     }
 
     /**
      * 返回集合中存在的key
-     * 
+     *
      * @param keys
      * @return
      */
@@ -156,7 +149,7 @@ public class RedisKeyUtil {
      * @param key
      */
     public boolean persistKey(String key) {
-        return redisTemplate.persist(key);
+        return Boolean.TRUE.equals(redisTemplate.persist(key));
     }
 
     /**
@@ -187,7 +180,7 @@ public class RedisKeyUtil {
      * @return 修改成功返回true
      */
     public boolean renameKeyNotExist(String oldKey, String newKey) {
-        return redisTemplate.renameIfAbsent(oldKey, newKey);
+        return Boolean.TRUE.equals(redisTemplate.renameIfAbsent(oldKey, newKey));
     }
 
     /**
@@ -196,20 +189,23 @@ public class RedisKeyUtil {
      * @param keys
      */
     public boolean deleteKey(Collection<String> keys) {
-        return redisTemplate.delete(keys) == keys.size();
+        Long delete = delete(keys);
+        return Objects.equals((long) keys.size(), delete);
     }
 
     /**
      * redis的key
      * 可形式为：
      * 表名:主键名:主键值:列名
-     * 
+     *
+     * @param split
+     * @param key
      * @return
      */
-    public static String getKey(String... key) {
+    public static String getKey(String split, String... key) {
         StringBuilder builder = new StringBuilder();
         for (String s : key) {
-            builder.append(s).append(":");
+            builder.append(s).append(split);
         }
         String s = builder.toString();
         return s.substring(0, s.length() - 1);
